[Template] Rename cart to shopping list
