+++ lib/kits/core-ui/components/paged-publication/index.js
@@ -1,5 +1,5 @@
 export {default as MainContainer} from './main-container';
 export {default as OfferList} from './offer-list';
-export {default as CartList} from './cart-list';
+export {default as ShoppingList} from './shopping-list';
 export {default as PageList} from './page-list';
 export {default as PublicationDownload} from './publication-download';
