+++ lib/kits/core-ui/components/paged-publication/main-container.js
@@ -7,7 +7,7 @@ const defaultTemplate = `\
             <div class="sgn__nav-logo"></div>
             <div class="sgn__nav">
                 <button class="sgn__offer-list">Offers</button>
-                <button class="sgn__offer-cart">List</button>
+                <button class="sgn__offer-shopping">Shopping List</button>
                 <button class="sgn__offer-pages">Pages</button>
                 <button class="sgn__offer-download">Download</button>
             </div>
