+++ lib/kits/core-ui/components/paged-publication/shopping-list.js
@@ -1,20 +1,20 @@
 import Mustache from 'mustache';
 import * as clientLocalStorage from '../../../../storage/client-local';
-import './cart-list.styl';
+import './shopping-list.styl';
 
 const defaultTemplate = `\
-    <div class="sgn-popup-contents sgn-cart-popup">
-        <span>List</span>
-        <ol class="sgn-cart-ol">
+    <div class="sgn-shopping-popup">
+        <span>Shopping List</span>
+        <ol class="sgn-shopping-ol">
             {{#offers}}
-            <li class="sgn-cart-li">
-                <div class="sgn-cart-li-flex-container">
-                    <div class="sgn-cart-li-div-text">
+            <li class="sgn-shopping-li">
+                <div class="sgn-shopping-li-flex-container">
+                    <div class="sgn-shopping-li-div-text">
                         <div class="sgn-truncate-elipsis">
                             <span>{{heading}}</span>
                         </div>
                     </div>
-                    <div class="sgn-cart-li-div-btn sgn-hide-print">
+                    <div class="sgn-shopping-li-div-btn sgn-hide-print">
                         <div>
                         </div>
                     </div>
@@ -23,8 +23,7 @@ const defaultTemplate = `\
             {{/offers}}
         </ol>
         <div>
-            <button class="sgn-cart-clear-cookie sgn-hide-print">Remove All</button>
-            <button class="sgn-cart-print sgn-hide-print" onclick="printCart()">Print</button>
+            <button class="sgn-shopping-clear-cookie sgn-hide-print">Remove All</button>
         </div>
     </div>\
 `;
@@ -32,14 +31,14 @@ const defaultTemplate = `\
 export default class CartList {
     constructor({template}) {
         this.template = template?.innerHTML || defaultTemplate;
-        this.cartListBtn = document.querySelector('.sgn__offer-cart');
+        this.shoppingListBtn = document.querySelector('.sgn__offer-shopping');
         this.headNav = document.querySelector('.sgn__nav');
         this.container = null;
         this.shown = false;
     }
 
     render() {
-        this.cartListBtn?.addEventListener(
+        this.shoppingListBtn?.addEventListener(
             'click',
             this.showCartList.bind(this),
             false
@@ -52,7 +51,7 @@ export default class CartList {
                 'paged-publication-saved-offers'
             );
             this.container = document.createElement('div');
-            this.container.className = 'sgn-cart-container';
+            this.container.className = 'sgn-shopping-container';
             this.insertAfter(this.headNav, this.container);
 
             this.container.innerHTML = Mustache.render(this.template, {
