+++ lib/kits/core-ui/components/paged-publication/shopping-list.styl
@@ -1,10 +1,11 @@
-.sgn-cart-popup
+.sgn-shopping-popup
     position fixed
     background #fff
     border-radius 6px
     padding 10px 0px 10px 10px
     top 50vh
     left 50vw
+    color #000000
     transform translate(-50%, -70%)
 
     > button
@@ -14,7 +15,7 @@
         margin 4px 2px
         cursor pointer
 
-.sgn-cart-dd
+.sgn-shopping-dd
     display block
     z-index 9
     position absolute
@@ -22,7 +23,7 @@
     /* border-radius 6px
     padding 10px 0px 10px 10px */
 
-.sgn-cart-ol
+.sgn-shopping-ol
     height 350px
     width 600px
     overflow-y scroll
@@ -30,7 +31,7 @@
     padding 10px
     color #000000
 
-.sgn-cart-li-flex-container
+.sgn-shopping-li-flex-container
     border-bottom 1px solid #eee
     height 40px
     align-items center
@@ -42,7 +43,7 @@
         margin 4px 2px
         cursor pointer
 
-.sgn-cart-li-div-text
+.sgn-shopping-li-div-text
     white-space nowrap
     overflow hidden
     text-overflow ellipsis
