+++ lib/kits/core-ui/paged-publication.js
@@ -25,7 +25,7 @@ export default class PagedPublication {
         this.start();
 
         this.renderOfferList();
-        this.renderCartList();
+        this.renderShoppingList();
         this.renderPageList();
         this.renderDownload();
     }
@@ -56,14 +56,14 @@ export default class PagedPublication {
         offerList.render();
     }
 
-    renderCartList() {
+    renderShoppingList() {
         const customTemplate = document.getElementById(
-            'sgn-sdk-paged-publication-viewer-cart-list-template'
+            'sgn-sdk-paged-publication-viewer-shopping-list-template'
         );
-        const cartList = new PagedPublicationComponents.CartList({
+        const shoppingList = new PagedPublicationComponents.ShoppingList({
             template: customTemplate
         });
-        cartList.render();
+        shoppingList.render();
     }
 
     renderPageList() {
